requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.37.0
bottle==0.12.25
bottle-websocket==0.2.9
customtkinter==5.2.0
darkdetect==0.8.0
Eel==0.16.0
et-xmlfile==1.1.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
speedtest-cli==2.1.3
tzdata==2023.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
